Move existing tests to the `serializer` package
