+++ data/src/test/java/ch/srg/dataProvider/integrationlayer/data/serializer/TestBlockReasonSerializer.kt
@@ -1,5 +1,6 @@
-package ch.srg.dataProvider.integrationlayer.data
+package ch.srg.dataProvider.integrationlayer.data.serializer
 
+import ch.srg.dataProvider.integrationlayer.data.DataProviderJson
 import ch.srg.dataProvider.integrationlayer.data.remote.BlockReason
 import kotlinx.serialization.encodeToString
 import kotlin.test.Test
