+++ data/src/test/java/ch/srg/dataProvider/integrationlayer/data/serializer/TestImageUrlSerializer.kt
@@ -1,5 +1,7 @@
-package ch.srg.dataProvider.integrationlayer.data
+package ch.srg.dataProvider.integrationlayer.data.serializer
 
+import ch.srg.dataProvider.integrationlayer.data.DataProviderJson
+import ch.srg.dataProvider.integrationlayer.data.ImageUrl
 import kotlinx.serialization.encodeToString
 import kotlin.test.Test
 import kotlin.test.assertEquals
